Partially implement result and input removal
New checks work for digit buttons and sign buttons
after the result is acquired.
There is an issue with the work of "remove"
button: for some reason there is a space appearing
in output after removal of a digit.

I would like to rework the logic because the
code is very confused and difficult to operate.
Intertwined method-based checks are a bad idea
after all.